hardware/detect_smell/requirements.txt
spidev
RPi.GPIO
requests
numpy
dht11
typing-extensions

